move mmdb to component
