+++ dns/filters.go
@@ -3,7 +3,7 @@ package dns
 import (
 	"net"
 
-	"github.com/Dreamacro/clash/common/mmdb"
+	"github.com/Dreamacro/clash/component/mmdb"
 )
 
 type fallbackFilter interface {
